+++ CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/appcelerator/titanium-webpack-devkit/compare/v0.2.3...v1.0.0) (2019-10-18)
+
+
+### Features
+
+* platform resolver plugin ([#25](https://github.com/appcelerator/titanium-webpack-devkit/issues/25)) ([de034c5](https://github.com/appcelerator/titanium-webpack-devkit/commit/de034c5))
+* titanium externals plugin ([#27](https://github.com/appcelerator/titanium-webpack-devkit/issues/27)) ([c74a694](https://github.com/appcelerator/titanium-webpack-devkit/commit/c74a694))
+
+
+### BREAKING CHANGES
+
+* replace the file system override with a resolver plugin
+
+
+
+
+
+## [0.2.3](https://github.com/appcelerator/titanium-webpack-devkit/compare/v0.2.2...v0.2.3) (2019-10-17)
+
+
+### Bug Fixes
+
+* replace NodeSourcePlugin with NodeTargetPlugin ([#23](https://github.com/appcelerator/titanium-webpack-devkit/issues/23)) ([341333d](https://github.com/appcelerator/titanium-webpack-devkit/commit/341333d))
+
+
+
+
+
+## [0.2.2](https://github.com/appcelerator/titanium-webpack-devkit/compare/v0.2.1...v0.2.2) (2019-07-01)
+
+
+### Bug Fixes
+
+* reference proper ipc message data ([#16](https://github.com/appcelerator/titanium-webpack-devkit/issues/16)) ([95838ba](https://github.com/appcelerator/titanium-webpack-devkit/commit/95838ba))
+
+
+
+
+
+## [0.2.1](https://github.com/appcelerator/titanium-webpack-devkit/compare/v0.1.0...v0.2.1) (2019-03-20)
+
+
+### Bug Fixes
+
+* **angular:** fix typo in require ([ac0493f](https://github.com/appcelerator/titanium-webpack-devkit/commit/ac0493f))
+
+
+
+
+
 <a name="0.1.0"></a>
 # [0.1.0](https://github.com/appcelerator/titanium-webpack-devkit/compare/v0.0.2...v0.1.0) (2018-06-14)